refactor: move theme into settings class
