+++ src/components/elements/ThemePicker.svelte
@@ -1,20 +1,22 @@
 <script lang="ts">
-	import activeTheme, { Theme } from '@stores/activeTheme';
+	import activeSettings, { Theme } from '@stores/activeSettings';
 	import { onMount } from 'svelte';
 	import { SunIcon, MoonIcon } from 'svelte-feather-icons';
 
 	onMount(() => {
 		// change theme if user prefers light color scheme
 		if (window.matchMedia('(prefers-color-scheme: light)').matches)
-			$activeTheme = Theme.light;
+			activeSettings.updateSelf(s => (s.theme = Theme.light));
 	});
 
 	const toggleTheme = () =>
-		activeTheme.update(x => (x === Theme.dark ? Theme.light : Theme.dark));
+		activeSettings.updateSelf(
+			s => (s.theme = s.theme === Theme.dark ? Theme.light : Theme.dark)
+		);
 </script>
 
 <button on:click={toggleTheme}>
-	{#if $activeTheme === Theme.light}
+	{#if $activeSettings.theme === Theme.light}
 		<MoonIcon />
 	{:else}
 		<SunIcon />
