+++ src/lib/settings.ts
@@ -1,0 +1,8 @@
+export enum Theme {
+	dark,
+	light,
+}
+
+export default class Settings {
+	constructor(public theme = Theme.dark) {}
+}
