+++ src/stores/activeSettings.ts
@@ -1,0 +1,6 @@
+import Settings from '@lib/settings';
+import { createWriteableObjectStore } from './createStore';
+export { Theme } from '@lib/settings';
+
+const activeSettings = createWriteableObjectStore(new Settings());
+export default activeSettings;
