--- src/stores/activeTheme.ts
@@ -1,8 +1,0 @@
-import { writable } from 'svelte/store';
-
-export enum Theme {
-	dark,
-	light,
-}
-
-export default writable<Theme>(Theme.dark);
